- Update slf4j dependencies  - Fix problem at ExecutorServiceFactory when initializing to big queue that produce OOM

+++ src/main/java/com/fnklabs/draenei/ExecutorServiceFactory.java
@@ -15,7 +15,7 @@ public class ExecutorServiceFactory {
             MAX_POOL_SIZE,
             0L,
             TimeUnit.SECONDS,
-            new ArrayBlockingQueue<Runnable>(Integer.MAX_VALUE),
+            new ArrayBlockingQueue<Runnable>(Short.MAX_VALUE),
             new ThreadFactory("draenei-executor-"),
             new ThreadPoolExecutor.CallerRunsPolicy()
     );
